% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nj.tree.R
\name{nj.tree}
\alias{nj.tree}
\title{Draw ggtree based phylogenetic tree using DNA or protein sequences}
\usage{
nj.tree(file, align.method = "muscle", html = TRUE, plot = TRUE)
}
\arguments{
\item{file}{a file path. The file containg the sequences in fasta format.}

\item{align.method}{method for multiple sequences alignment, one of "muscle" and "DECIPHER"}

\item{html}{a boolean value to indicate if to show the alignment result in the explore}

\item{plot}{a boolean value if draw the tree using ggtree package}
}
\value{
a ggtree plot
}
\description{
Draw ggtree based phylogenetic tree using DNA or protein sequences
}
\examples{
nj.tree("prob.16S.fa")
}
\author{
Zhonghui Gai
}
